+++ litsinger.ernest/P4/function.hpp
@@ -5,6 +5,5 @@
 namespace litsinger
 {
   size_t inputArray(std::ifstream & input, int * matrix, size_t sizeMatrix);
-  int MaximalSum(const int * arr, size_t rows, size_t cols);
 }
 #endif
