+++ litsinger.ernest/P4/variantlogic.hpp
@@ -1,0 +1,7 @@
+#ifndef VARIANTLOGIC_HPP
+#define VARIANTLOGIC_HPP
+namespace litsinger
+{
+  int MaximalSum(const int* arr, size_t rows, size_t cols);
+}
+#endif
